+++ vllm/model_executor/layers/fused_moe/layer.py
@@ -10,8 +10,8 @@ from vllm.logger import init_logger
 from vllm.model_executor.layers.fused_moe.fused_moe import fused_moe
 from vllm.model_executor.layers.quantization.base_config import (
     QuantizationConfig, QuantizeMethodBase)
-from vllm.model_executor.model_loader.deferred_tensor import convert_like
 from vllm.model_executor.utils import set_weight_attrs
+from vllm.utils import convert_like
 
 logger = init_logger(__name__)
